[PVR] Refactor CPVRGUIActions step 7: Move parental control related actions to own class.
